+++ R/cdm.R
@@ -6,13 +6,6 @@
 #'   to NULL.
 #' @param write_schema,writeSchema An optional schema in the CDM database that the user has
 #'   write access to.
-#' @param cdm_tables,cdmTables Which tables should be included? Supports a character
-#'   vector, tidyselect selection helpers, or table groups.
-#' \itemize{
-#'   \item{tbl_group("all")}{all CDM tables}
-#'   \item{tbl_group("vocab")}{the CDM vocabulary tables}
-#'   \item{tbl_group("clinical")}{the clinical CDM tables}
-#' }
 #' @param cohort_tables,cohortTables A character vector listing the cohort table names to be
 #'   included in the CDM object.
 #' @param cdm_version,cdmVersion The version of the OMOP CDM: "5.3" (default), "5.4",
@@ -20,20 +13,18 @@
 #'   heuristics. Cohort tables must be in the write_schema.
 #' @param cdm_name,cdmName The name of the CDM. If NULL (default) the cdm_source_name
 #'.  field in the CDM_SOURCE table will be used.
-#' @param write_prefix,writePrefix A prefix that should be used with all tables
-#' written to the "write_schema". This prefix allows for the creation of a
-#' namespace within a database schema.
+#'
 #' @return A list of dplyr database table references pointing to CDM tables
 #' @importFrom dplyr all_of matches starts_with ends_with contains
 #' @export
 cdm_from_con <- function(con,
                          cdm_schema = NULL,
-                         cdm_tables = tbl_group("default"),
                          write_schema = NULL,
                          cohort_tables = NULL,
                          cdm_version = "5.3",
-                         cdm_name = NULL,
-                         write_prefix = NULL) {
+                         cdm_name = NULL) {
+
+  cdm_tables <- tbl_group("all")
 
   checkmate::assert_class(con, "DBIConnection")
   checkmate::assert_true(.dbIsValid(con))
@@ -42,43 +33,16 @@ cdm_from_con <- function(con,
     cdm_schema = "main"
   }
 
-  checkmate::assert_character(
-    cdm_schema,
-    null.ok = FALSE,
-    min.len = 1,
-    max.len = 2
-  )
-  checkmate::assert_character(
-    write_schema,
-    null.ok = TRUE,
-    min.len = 1,
-    max.len = 2
-  )
+  cdm_schema_list <- normalize_schema(cdm_schema)
+  cdm_schema <- cdm_schema_list$schema
+  cdm_prefix <- cdm_schema_list$prefix
+  write_schema_list <- normalize_schema(write_schema)
+  write_schema <- write_schema_list$schema
+  write_prefix <- write_schema_list$prefix
+
   checkmate::assert_character(cohort_tables, null.ok = TRUE, min.len = 1)
   checkmate::assert_choice(cdm_version, choices = c("5.3", "5.4", "auto"))
   checkmate::assert_character(cdm_name, null.ok = TRUE)
-  checkmate::assert_character(write_prefix, null.ok = TRUE, len = 1, min.chars = 1, pattern = "[_a-z]+")
-
-  # handle schema names like 'schema.dbo'
-  if (!is.null(cdm_schema) && length(cdm_schema) == 1) {
-    cdm_schema <- strsplit(cdm_schema, "\\.")[[1]]
-      checkmate::assert_character(
-        cdm_schema,
-        null.ok = TRUE,
-        min.len = 1,
-        max.len = 2
-      )
-    }
-
-  if (!is.null(write_schema) && length(write_schema) == 1) {
-    write_schema <- strsplit(write_schema, "\\.")[[1]]
-    checkmate::assert_character(
-      write_schema,
-      null.ok = TRUE,
-      min.len = 1,
-      max.len = 2
-    )
-  }
 
   if (cdm_version == "auto") {
     cdm_version <- detect_cdm_version(con, cdm_schema = cdm_schema)
@@ -102,93 +66,130 @@ cdm_from_con <- function(con,
                                 cdm_source$cdm_source_abbreviation[1])
   }
 
-    # tidyselect: https://tidyselect.r-lib.org/articles/tidyselect.html
-    all_cdm_tables <-
-      rlang::set_names(spec_cdm_table[[cdm_version]]$cdmTableName,
-                       spec_cdm_table[[cdm_version]]$cdmTableName)
-
-    cdm_tables <- names(tidyselect::eval_select(rlang::enquo(cdm_tables),
-                                                data = all_cdm_tables))
-
-    # Handle uppercase table names in the database
-    if (all(dbTables == toupper(dbTables))) {
-      cdm_tables <- toupper(cdm_tables)
-    }
-
-    cdm <- purrr::map(cdm_tables, ~dplyr::tbl(con, inSchema(cdm_schema, ., dbms(con))) %>%
-                      dplyr::rename_all(tolower)) %>%
-      rlang::set_names(tolower(cdm_tables))
-
-    if (!is.null(write_schema)) {
-      verify_write_access(con, write_schema = write_schema)
-    }
-
-    dbTablesWrite <- listTables(con, schema = write_schema)
-    # Add existing GeneratedCohortSet objects to cdm object
-    if (!is.null(cohort_tables)) {
-      if (is.null(write_schema)) {
-        rlang::abort("write_schema is required when using cohort_tables")
-      }
-
-      for (i in seq_along(cohort_tables)) {
-
-        cohort_ref <- dplyr::tbl(con, inSchema(write_schema, cohort_tables[i], dbms(con))) %>%
-          dplyr::rename_all(tolower)
-
-        # Optional attribute tables {cohort}_set, {chohort}_attrition, {cohort}_count
-        nm <- paste0(cohort_tables[i], "_set")
-        if (nm %in% dbTablesWrite) {
-          cohort_set_ref <- dplyr::tbl(con, inSchema(write_schema, nm, dbms(con))) %>%
-            dplyr::rename_all(tolower)
-        } else if (nm %in% toupper(dbTablesWrite)) {
-          cohort_set_ref <- dplyr::tbl(con, inSchema(write_schema, toupper(nm), dbms(con))) %>%
-            dplyr::rename_all(tolower)
-        } else {
-          cohort_set_ref <- NULL
-        }
-
-        nm <- paste0(cohort_tables[i], "_attrition")
-        if (nm %in% dbTablesWrite) {
-          cohort_attrition_ref <- dplyr::tbl(con, inSchema(write_schema, nm, dbms(con))) %>%
-            dplyr::rename_all(tolower)
-        } else if (nm %in% toupper(dbTablesWrite)) {
-          cohort_attrition_ref <- dplyr::tbl(con, inSchema(write_schema, toupper(nm), dbms(con))) %>%
-            dplyr::rename_all(tolower)
-        } else {
-          cohort_attrition_ref <- NULL
-        }
-
-        nm <- paste0(cohort_tables[i], "_count")
-        if (nm %in% dbTablesWrite) {
-          cohort_count_ref <- dplyr::tbl(con, inSchema(write_schema, nm, dbms(con))) %>%
-            dplyr::rename_all(tolower)
-        } else if (nm %in% toupper(dbTablesWrite)) {
-          cohort_count_ref <- dplyr::tbl(con, inSchema(write_schema, toupper(nm), dbms(con))) %>%
-            dplyr::rename_all(tolower)
-        } else {
-          cohort_count_ref <- NULL
-        }
-
-
-        cdm[[cohort_tables[i]]] <- new_generated_cohort_set(
-          cohort_ref = cohort_ref,
-          cohort_attrition_ref = cohort_attrition_ref,
-          cohort_set_ref = cohort_set_ref,
-          cohort_count_ref = cohort_count_ref)
-      }
-    }
-
-
-    # TODO cdm_reference constructor
-    class(cdm) <- "cdm_reference"
-    attr(cdm, "cdm_schema") <- cdm_schema
-    attr(cdm, "write_schema") <- write_schema
-    attr(cdm, "write_prefix") <- write_prefix
-    attr(cdm, "dbcon") <- con
-    attr(cdm, "cdm_version") <- cdm_version
-    attr(cdm, "cdm_name") <- cdm_name
-    return(cdm)
+  # only get the cdm tables that exist in the database
+  cdm_tables <- cdm_tables[which(cdm_tables %in% tolower(dbTables))]
+  if (length(cdm_tables) == 0) {
+    rlang::abort("There were no cdm tables found in the cdm_schema!")
   }
+
+  # Add prefix if supplied. If not supplied cdm_prefix will be NULL
+  cdm_tables_prefixed <- paste0(cdm_prefix, cdm_tables)
+
+  # Handle uppercase table names in the database
+  if (all(dbTables == toupper(dbTables))) {
+    cdm_tables_prefixed <- toupper(cdm_tables_prefixed)
+  }
+
+  cdm <- purrr::map(cdm_tables_prefixed, ~dplyr::tbl(con, inSchema(cdm_schema, ., dbms(con))) %>%
+                    dplyr::rename_all(tolower)) %>%
+    rlang::set_names(cdm_tables)
+
+  if (!is.null(write_schema)) {
+    verify_write_access(con, write_schema = write_schema)
+  }
+
+  write_schema_tables <- listTables(con, schema = write_schema)
+  # Add existing GeneratedCohortSet objects to cdm object
+  if (!is.null(cohort_tables)) {
+    if (is.null(write_schema)) {
+      rlang::abort("write_schema is required when using cohort_tables")
+    }
+
+    for (i in seq_along(cohort_tables)) {
+
+      cohort_table <- paste0(write_prefix, cohort_tables[i])
+
+      # A generated cohort set object has tables: {cohort}
+      # and attribute tables {cohort}_set, {cohort}_attrition, {cohort}_count
+      # Only {cohort}_attrition is optional. The others will be created if not supplied.
+      nms <- paste0(cohort_table, c("", "_set", "_count", "_attrition"))
+      x <- purrr::map(nms, function(nm) {
+        if (nm %in% write_schema_tables) {
+          dplyr::tbl(con, inSchema(write_schema, nm, dbms(con))) %>%
+            dplyr::rename_all(tolower)
+        } else if (nm %in% toupper(write_schema_tables)) {
+          dplyr::tbl(con, inSchema(write_schema, toupper(nm), dbms(con))) %>%
+            dplyr::rename_all(tolower)
+        } else {
+          NULL
+        }
+      })
+
+      cohort_ref <- x[[1]]
+      cohort_set_ref <- x[[2]]
+      cohort_count_ref <- x[[3]]
+      cohort_attrition_ref <- x[[4]]
+
+      if (is.null(cohort_ref)) {
+        rlang::abort(glue::glue("cohort table `{cohort_table}` not found!"))
+      }
+
+      if (is.null(cohort_set_ref)) {
+        # create the required cohort_set table
+        cohort_set_ref <- cohort_ref %>%
+          dplyr::distinct(.data$cohort_definition_id) %>%
+          dplyr::mutate(cohort_name = paste("cohort", .data$cohort_definition_id)) %>%
+          computeQuery(name = paste0(cohort_table, "_set"),
+                       schema = write_schema,
+                       temporary = FALSE,
+                       overwrite = TRUE)
+      }
+
+      if (is.null(cohort_count_ref)) {
+        # create the required cohort_count table
+        cohort_count_ref <- cohort_ref %>%
+          dplyr::ungroup() %>%
+          dplyr::group_by(.data$cohort_definition_id) %>%
+          dplyr::summarise(number_records = dplyr::n(),
+                           number_subjects = dplyr::n_distinct(.data$subject_id)) %>%
+          computeQuery(name = paste0(cohort_table, "_count"),
+                       schema = write_schema,
+                       temporary = FALSE,
+                       overwrite = TRUE)
+      }
+
+      # Note: use name without prefix (i.e. `cohort_tables[i]`) in the cdm object
+      cdm[[cohort_tables[i]]] <- new_generated_cohort_set(
+        cohort_ref = cohort_ref,
+        cohort_set_ref = cohort_set_ref,
+        cohort_count_ref = cohort_count_ref,
+        cohort_attrition_ref = cohort_attrition_ref)
+    }
+  }
+
+  # TODO use a cdm_reference constructor function
+  class(cdm) <- "cdm_reference"
+  attr(cdm, "cdm_schema") <- cdm_schema
+  attr(cdm, "write_schema") <- write_schema
+  attr(cdm, "write_prefix") <- write_prefix
+  attr(cdm, "cdm_prefix") <- cdm_prefix
+  attr(cdm, "dbcon") <- con
+  attr(cdm, "cdm_version") <- cdm_version
+  attr(cdm, "cdm_name") <- cdm_name
+  # The following attributes can be used to communicate temp table preferences to downstream analytic packages.
+  # This a feature for analytic package developers and users should not need to know about it unless there is an issue to debug.
+  attr(cdm, "cohort_as_temp") <- FALSE
+  attr(cdm, "intermediate_as_temp") <- TRUE
+  return(cdm)
+}
+
+#' @rdname cdm_from_con
+#' @export
+cdmFromCon <- function(con,
+                       cdmSchema = NULL,
+                       writeSchema = NULL,
+                       cohortTables = NULL,
+                       cdmVersion = "5.3",
+                       cdmName = NULL) {
+  cdm_from_con(
+    con = con,
+    cdm_schema = cdmSchema,
+    write_schema = writeSchema,
+    cohort_tables = cohortTables,
+    cdm_version = cdmVersion,
+    cdm_name = cdmName
+  )
+}
 
 detect_cdm_version <- function(con, cdm_schema = NULL) {
   cdm_tables <- c("visit_occurrence", "cdm_source", "procedure_occurrence")
@@ -197,7 +198,7 @@ detect_cdm_version <- function(con, cdm_schema = NULL) {
     rlang::abort(paste0(
       "The ",
       paste(cdm_tables, collapse = ", "),
-      " tables are require for auto-detection of cdm version."
+      " tables are required for auto-detection of cdm version."
     ))
   }
 
@@ -220,7 +221,6 @@ detect_cdm_version <- function(con, cdm_schema = NULL) {
     return("5.4")
   }
 
-
   procedure_occurrence_names <- cdm$procedure_occurrence %>%
     head() %>%
     collect() %>%
@@ -232,15 +232,15 @@ detect_cdm_version <- function(con, cdm_schema = NULL) {
   }
 
   cdm_version <- cdm$cdm_source %>% dplyr::pull(.data$cdm_version)
-  if (isTRUE(grepl("5\\.4", cdm_version)))
-    return("5.4")
-  if (isTRUE(grepl("5\\.3", cdm_version)))
-    return("5.3")
+  if (isTRUE(grepl("5\\.4", cdm_version))) return("5.4")
 
-  if ("episode" %in% listTables(con, schema = cdm_schema))
+  if (isTRUE(grepl("5\\.3", cdm_version))) return("5.3")
+
+  if ("episode" %in% listTables(con, schema = cdm_schema)) {
     return("5.4")
-  else
+  } else {
     return("5.3")
+  }
 }
 
 #' Get the CDM version
@@ -256,7 +256,7 @@ detect_cdm_version <- function(con, cdm_schema = NULL) {
 #' \dontrun{
 #' library(CDMConnector)
 #' con <- DBI::dbConnect(duckdb::duckdb(), eunomia_dir())
-#' cdm <- cdm_from_con(con, cdm_tables = c(tbl_group("default"), "cdm_source"))
+#' cdm <- cdm_from_con(con, "main")
 #' version(cdm)
 #'
 #' DBI::dbDisconnect(con, shutdown = TRUE)
@@ -304,30 +304,6 @@ cdmName <- function(cdm) {
 #' @export
 cdm_name <- cdmName
 
-
-#' @rdname cdm_from_con
-#' @export
-cdmFromCon <- function(con,
-                       cdmSchema = NULL,
-                       cdmTables = tbl_group("default"),
-                       writeSchema = NULL,
-                       cohortTables = NULL,
-                       cdmVersion = "5.3",
-                       cdmName = NULL,
-                       writePrefix = NULL) {
-  cdm_from_con(
-    con = con,
-    cdm_schema = cdmSchema,
-    cdm_tables = {{cdmTables}},
-    write_schema = writeSchema,
-    cohort_tables = cohortTables,
-    cdm_version = cdmVersion,
-    cdm_name = cdmName,
-    write_prefix = writePrefix
-  )
-}
-
-
 #' Print a CDM reference object
 #'
 #' @param x A cdm_reference object
@@ -358,12 +334,11 @@ verify_write_access <- function(con, write_schema, add = NULL) {
 
   # TODO quote SQL names
   write_schema <- paste(write_schema, collapse = ".")
-  tablename <-
-    paste(c(sample(letters, 12, replace = TRUE), "_test_table"), collapse = "")
+  tablename <- paste(c(sample(letters, 12, replace = TRUE), "_test_table"), collapse = "")
   tablename <- paste(write_schema, tablename, sep = ".")
 
   df1 <- data.frame(chr_col = "a", numeric_col = 1)
-  # ROracle does not support integer round trip
+  # Note: ROracle does not support integer round trip
   DBI::dbWriteTable(con, DBI::SQL(tablename), df1)
 
   withr::with_options(list(databaseConnectorIntegerAsNumeric = FALSE), {
@@ -374,9 +349,7 @@ verify_write_access <- function(con, write_schema, add = NULL) {
   DBI::dbRemoveTable(con, DBI::SQL(tablename))
 
   if (!isTRUE(all.equal(df1, df2))) {
-    msg <- paste("Write access to schema",
-                 write_schema,
-                 "could not be verified.")
+    msg <- paste("Write access to schema", write_schema, "could not be verified.")
 
     if (is.null(add)) {
       rlang::abort(msg)
@@ -417,7 +390,8 @@ verify_write_access <- function(con, write_schema, add = NULL) {
 #'                       user = "postgres",
 #'                       password = Sys.getenv("PASSWORD"))
 #'
-#' cdm <- cdm_from_con(con, cdm_tables = tbl_group("vocab"))
+#' cdm <- cdm_from_con(con) %>%
+#'   cdm_select_tbl(tbl_group("vocab"))
 #' }
 tbl_group <- function(group) {
   # groups are defined in the internal package dataframe called spec_cdm_table
@@ -489,7 +463,7 @@ dbms.DBIConnection <- function(con) {
 #'
 #' @param cdm A cdm object
 #' @param path A folder to save the cdm object to
-#' @param format The file format to use: "parquet", "csv", "feather", "duckdb".
+#' @param format The file format to use: "parquet" (default), "csv", "feather" or "duckdb".
 #'
 #' @return Invisibly returns the cdm input
 #' @export
@@ -497,7 +471,8 @@ dbms.DBIConnection <- function(con) {
 #' @examples
 #' \dontrun{
 #' con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
-#' vocab <- cdm_from_con(con, cdm_tables = c("concept", "concept_ancestor"))
+#' vocab <- cdm_from_con(con, "main") %>%
+#'   cdm_select_tbl("concept", "concept_ancestor")
 #' stow(vocab, here::here("vocab_tables"))
 #' DBI::dbDisconnect(con, shutdown = TRUE)
 #' }
@@ -527,7 +502,7 @@ stow <- function(cdm, path, format = "parquet") {
     duckdb = {
       rlang::check_installed("duckdb")
       con <- DBI::dbConnect(duckdb::duckdb(), dbdir = file.path(path, "cdm.duckdb"))
-      purrr::walk2(cdm, names(cdm), ~DBI::dbWriteTable(con, name = dplyr::collect(.x), value = .y))
+      purrr::walk(names(cdm), ~DBI::dbWriteTable(con, name = ., value = dplyr::collect(cdm[[.]])))
       DBI::dbDisconnect(con, shutdown = TRUE)
     }
   )
@@ -544,60 +519,56 @@ stow <- function(cdm, path, format = "parquet") {
 #'   FALSE will read files into R as Arrow Datasets.
 #' @return A list of dplyr database table references pointing to CDM tables
 #' @export
-cdm_from_files <-
-  function(path,
-           format = "auto",
-           as_data_frame = TRUE) {
-    checkmate::assert_choice(format, c("auto", "parquet", "csv", "feather"))
-    checkmate::assert_logical(as_data_frame, len = 1, null.ok = FALSE)
-    checkmate::assert_true(file.exists(path))
+cdm_from_files <- function(path,
+                           format = "auto",
+                           as_data_frame = TRUE) {
+  checkmate::assert_choice(format, c("auto", "parquet", "csv", "feather"))
+  checkmate::assert_logical(as_data_frame, len = 1, null.ok = FALSE)
+  checkmate::assert_true(file.exists(path))
 
-    path <- path.expand(path)
+  path <- path.expand(path)
 
-    files <- list.files(path, full.names = TRUE)
+  files <- list.files(path, full.names = TRUE)
 
-    if (format == "auto") {
-      format <- unique(tools::file_ext(files))
-      if (length(format) > 1)
-        rlang::abort(paste(
-          "Multiple file formats detected:",
-          paste(format, collapse = ", ")
-        ))
-      checkmate::assert_choice(format, c("parquet", "csv", "feather"))
+  if (format == "auto") {
+    format <- unique(tools::file_ext(files))
+    if (length(format) > 1) {
+      rlang::abort(paste("Multiple file formats detected:", paste(format, collapse = ", ")))
     }
-
-    cdm_tables <-
-      tools::file_path_sans_ext(basename(list.files(path)))
-
-    cdm_table_files <-
-      file.path(path, paste0(cdm_tables, ".", format))
-
-    purrr::walk(cdm_table_files, function(.) {
-      checkmate::assert_file_exists(., "r")
-    })
-
-    cdm <- switch(
-      format,
-      parquet = purrr::map(cdm_table_files, function(.) {
-        arrow::read_parquet(., as_data_frame = as_data_frame)
-      }),
-      csv = purrr::map(cdm_table_files, function(.) {
-        arrow::read_csv_arrow(., as_data_frame = as_data_frame)
-      }),
-      feather = purrr::map(cdm_table_files, function(.) {
-        arrow::read_feather(., as_data_frame = as_data_frame)
-      })
-    ) %>%
-      magrittr::set_names(cdm_tables) %>%
-      magrittr::set_class("cdm_reference")
-
-    attr(cdm, "cdm_schema") <- NULL
-    attr(cdm, "write_schema") <- NULL
-    attr(cdm, "dbcon") <- NULL
-    attr(cdm, "cdm_version") <- NULL
-    attr(cdm, "write_prefix") <- NULL
-    cdm
+    checkmate::assert_choice(format, c("parquet", "csv", "feather"))
   }
+
+  cdm_tables <- tools::file_path_sans_ext(basename(list.files(path)))
+  cdm_table_files <- file.path(path, paste0(cdm_tables, ".", format))
+  purrr::walk(cdm_table_files, ~checkmate::assert_file_exists(., "r"))
+
+  cdm <- switch(
+    format,
+    parquet = purrr::map(cdm_table_files, function(.) {
+      arrow::read_parquet(., as_data_frame = as_data_frame)
+    }),
+    csv = purrr::map(cdm_table_files, function(.) {
+      arrow::read_csv_arrow(., as_data_frame = as_data_frame)
+    }),
+    feather = purrr::map(cdm_table_files, function(.) {
+      arrow::read_feather(., as_data_frame = as_data_frame)
+    })
+  ) %>%
+    magrittr::set_names(cdm_tables) %>%
+    magrittr::set_class("cdm_reference")
+
+  attr(cdm, "cdm_schema") <- NULL
+  attr(cdm, "write_schema") <- NULL
+  attr(cdm, "write_prefix") <- NULL
+  attr(cdm, "cdm_prefix") <- NULL
+  attr(cdm, "dbcon") <- NULL
+  attr(cdm, "cdm_version") <- NULL
+  attr(cdm, "cdm_name") <- NULL
+  attr(cdm, "cohort_as_temp") <- NULL
+  attr(cdm, "intermediate_as_temp") <- NULL
+
+  cdm
+}
 
 
 #' @rdname cdm_from_files
@@ -626,7 +597,9 @@ cdmFromFiles <- function(path,
 #' @examples
 #' \dontrun{
 #' con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
-#' vocab <- cdm_from_con(con, cdm_tables = c("concept", "concept_ancestor"))
+#' vocab <- cdm_from_con(con, "main") %>%
+#'   cdm_select_tbl("concept", "concept_ancestor")
+#'
 #' local_vocab <- collect(vocab)
 #' DBI::dbDisconnect(con, shutdown = TRUE)
 #' }
@@ -653,11 +626,13 @@ NULL
 #' observation_period tables
 #' @export
 #'
+#' @importFrom tidyr gather
+#'
 #' @examples
 #' \dontrun{
 #' library(CDMConnector)
 #' con <- DBI::dbConnect(duckdb::duckdb(), eunomia_dir())
-#' cdm <- cdm_from_con(con, cdm_tables = c(tbl_group("default"), "cdm_source"))
+#' cdm <- cdm_from_con(con, "main")
 #' snapshot(cdm)
 #'
 #' DBI::dbDisconnect(con, shutdown = TRUE)
@@ -707,21 +682,13 @@ snapshot <- function(cdm) {
       "person_cnt",
       "observation_period_cnt"
     ) %>%
-    as.list() %>%
-    c(list(cdm_schema = attr(cdm, "cdm_schema"),
-           write_schema = attr(cdm, "write_schema"),
-           cdm_name = attr(cdm, "cdm_name"))) %>%
-    magrittr::set_class("cdm_snapshot")
-
-}
-
-#' @export
-print.cdm_snapshot <- function(x, ...) {
-  cli::cat_rule(x$cdm_source_name)
-  if (length(x$cdm_schema) > 1) {
-    x$cdm_schema <- paste0(x$cdm_schema, collapse = ".")
-  }
-  purrr::walk2(names(x[-1]), x[-1], ~ cli::cat_bullet(.x, ": ", .y))
+    dplyr::mutate(cdm_schema = attr(cdm, "cdm_schema"),
+                  write_schema = attr(cdm, "write_schema"),
+                  cdm_name = attr(cdm, "cdm_name")) %>%
+    dplyr::mutate_all(as.character) %>%
+    tidyr::gather(key = "attribute", value = "value") %>%
+    dplyr::mutate(value = ifelse(.data$value == "", NA_character_, .data$value)) %>%
+    dplyr::tibble()
 }
 
 #' Disconnect the connection of the cdm object